auth_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.7.0
pydantic[email]==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.0.1
pydantic-settings==2.6.1
email-validator==2.1.0
certifi==2024.2.2